feat: unify wasm modules structure
