+++ Makefile
@@ -22,16 +22,16 @@ LISTEN_ADDR := 127.0.0.1:8000
 # Repository URL for issues.
 REPO_URL := $(shell git remote get-url origin | sed -e 's/:/\//' -e 's/git@/https:\/\//' -e 's/\.git//')
 
-# Exports
-export REACT_APP_VERSION=$(APP_VERSION)
-export REACT_APP_WASM_API_VER=$(WASM_API_VER)
-export REACT_APP_GITHUB_URL=$(REPO_URL)
-
 .PHONY:all
 all: build
 
 include build.mk
 include docker.mk
+
+# Exports
+export REACT_APP_VERSION=$(APP_VERSION)
+export REACT_APP_WASM_API_VER=$(WASM_API_VER)
+export REACT_APP_GITHUB_URL=$(REPO_URL)
 
 .PHONY:run
 run:
