+++ cmd/README.md
@@ -1,0 +1,11 @@
+# cmd
+
+This directory contains buildable project binaries.
+
+## playground
+
+The playground is a web server that serves the playground.
+
+## wasm
+
+The wasm directory contains WebAssembly binaries used on frontend side.
