+++ cmd/wasm/analyzer/README.md
@@ -1,0 +1,6 @@
+# Analyzer
+
+`analyzer` binary exports collection of tools used by UI web worker to analyze Go code and report
+code errors.
+ 
+Package should be compiled as **WebAssembly** binary and loaded by JS WebWorker.
