+++ cmd/wasm/analyzer/webworker.go
