+++ cmd/wasm/api-version.txt
@@ -1,0 +1,1 @@
+v2
