--- cmd/webworker/README.md
@@ -1,6 +1,0 @@
-# WebWorker
-
-`webworker` binary exports collection of tools used by UI web worker to analyze Go code and report
-code errors.
- 
-Package should be compiled as **WebAssembly** binary and loaded by JS WebWorker.
